Undoing change committed in r12+13.
git-svn-id: svn://svn.gps.alaska.edu/crusde/trunk@15 1095105e-3e63-0410-b14e-63ce1eee999c

+++ src/InputHandler.cpp
@@ -583,19 +583,16 @@ string InputHandler::getResultFileName()
 	
 }
 
-double InputHandler::getRegion(string direction)
+int InputHandler::getRegion(string direction)
 {
 	StrXML d(direction);
-//	string s(getAttributeValueByName(TAG_region.xmlStr(), ATTR_value.xmlStr(), d.xmlStr()).c_str());
-//	crusde_debug("%s, line: %d, InputHandler::getRegion %s=%s=%f", __FILE__, __LINE__, direction.c_str(), s.c_str(), atof(s.c_str()));
-	
-	return atof(getAttributeValueByName(TAG_region.xmlStr(), ATTR_value.xmlStr(), d.xmlStr()).c_str());
+	return atoi(getAttributeValueByName(TAG_region.xmlStr(), ATTR_value.xmlStr(), d.xmlStr()).c_str());
 }
 
-double InputHandler::getGridSize()
+int InputHandler::getGridSize()
 {
 	StrXML g("gridsize");
-	return atof(getAttributeValueByName(TAG_parameter.xmlStr(), ATTR_value.xmlStr(), g.xmlStr()).c_str());
+	return atoi(getAttributeValueByName(TAG_parameter.xmlStr(), ATTR_value.xmlStr(), g.xmlStr()).c_str());
 }
 
 unsigned long int InputHandler::getTimeSteps()
